// frontend/src/types/types.ts
export interface IPost {
  id: string;
  privacy: "public" | "private" | "follower";
  photos: Record<string, any>[];
  comments: any[];
  description: string;
  likes: any[];
  author: IUser;
  commentsCount: number;
  likesCount: number;
  isBookmarked: boolean;
  isLiked: boolean;
  isOwnPost: boolean;
  createdAt: Date;
  updatedAt: Date;
}

export interface IUser {
  id: string;
  username: string;
  fullName: string;
  email: string;
  profilePicture?: string | Record<string, any>;
  isEmailValidated: boolean;
  [prop: string]: any;
}
// ...
export interface IProfile {
  _id: string;
  id: string;
  username: string;
  email: string;
  fullname: string;
  firstname: string;
  lastname: string;
  info: {
    bio: string;
    birthday: string;
    gender: string;
  };
  isEmailValidated: boolean;
  profilePicture?: string | Record<string, any>;
  isOwnProfile: boolean;
  coverPhoto?: string | Record<string, any>;
  dateJoined: Date | string;
  followingCount: number;
  followersCount: number;
  [prop: string]: any;
}

export interface IPostState {
  user_id: string;
  items: IPost[];
  offset: number;
}

export interface IUserState {
  data: IProfile | null;
  posts: IPost[];
  postsOffset: number;
  followers: IProfile[];
  followersOffset: number;
  following: IProfile[];
  followingOffset: number;
  bookmarks: IBookmark[];
  bookmarksOffset: number;
  developer: IProfile | null;
}
// ...
export interface IBookmark {
  id: string;
  isBookmarked: boolean;
  post: IPost;
  createdAt: Date;
}